Changed MaybeTopOfBorrowStack analysis
Now allowing immutable borrows to be above the tracked local on the
borrow stack. The analysis will consider the local to be on top as
long as it is the top mutable borrow.

Casts of references to immutable pointers are still treated the same as
casts to mutable pointers. (--> the analysis views a reference to not be
on top of the stack in that case.)